refactor: Moved constants to a constants package
