+++ internal/constants/bool.go
@@ -1,0 +1,6 @@
+package constants
+
+const (
+	Yes = "yes"
+	No  = "no"
+)
