+++ internal/constants/vendors.go
@@ -1,0 +1,7 @@
+package constants
+
+const (
+	VendorCrowdStrike = "CrowdStrike, Inc."
+	VendorRedHat      = "Red Hat, Inc."
+	VendorGNU         = "GNU"
+)
